improved RTE flat to correct color cast
Improved version of the RTE flat to correct for significant color cast that remained after calibration with the reverse engineered flat.
V2 calibrated RTE images were stacked and the gradients were modeled in PixInsight. This was combined with Flat V2.
The improvements were verified by creating a mosaic texture for a photogrammetry reconstruction of flight 10 that showed much smoother seam lines.